# ENH: help compilers to auto-vectorize reduction operators