--- a/doc/release/upcoming_changes/21001.performance.rst
+++ b/doc/release/upcoming_changes/21001.performance.rst
@@ -1,3 +1,5 @@
 Faster reduction operators
----------------------
-Reduction operations like `numpy.sum`, `numpy.prod`, `numpy.add.reduce`, `numpy.logical_and.reduce` on contiguous integer-based arrays are now much faster.
+--------------------------
+Reduction operations like `numpy.sum`, `numpy.prod`, `numpy.add.reduce`, 
+`numpy.logical_and.reduce` on contiguous integer-based arrays are now 
+much faster.