@seberg 2022-02-06
Hmm, had not noticed the test failure.  Might be another reason to split up the two, since I am not sure they come from the reductions or the cast change.  It looks like it should be due to the cast change, but it is not clear to me why.