@zephyr111 2022-02-06
> It looks like it should be due to the cast change, but it is not clear to me why.

I saw that too but I found this very weird since I did not really changed the actual computation, just some kind of compiler "hints". Your point about alignment make me thing about either GCC could assume that the input is aligned with AVX2 and simply generate a wrong code for this benchmark. That being said, it would be very surprising since AFAIK GCC should not make such an assumption unless explicitly stated. Otherwise, I guess it could mean the current code has an undefined behavior and using AVX2 make it visible due to different optimizations (which is scary). 

>  I suspect this may be undefined behaviour in C99.

It makes sense! This is indeed undefined behavior to cast a float value to an integer that cannot represent the integer part. See [this](https://stackoverflow.com/a/50134990/12939557) StackOverflow post.

> If you are up for it, I think it would make sense to split this into two PRs. I am fine with this, but how am I supposed to proceed? Should I just need to revert the change on this PR and open a new PR with this one mentioned in it so to merge this one before?