@seberg 2022-02-06
> I am fine with this, but how am I supposed to proceed? Should I just need to revert the change on this PR and open a new PR with this one mentioned in it so to merge this one before?

Sounds good, i.e. remove the casting part of the PR and create anew PR for that (or vice versa, I guess).  About the undefined behaviour, I have to think.  I am not surprised it behaves differently for UB, but if it was effectively always well defined, so we have to decide that this is a weird corner case that should not matter, or see if there is a trick to preserve behaviour.
(We could be pragmatic and split out vectorizing those casts, because unsafe casts are less common anyway.)