@zephyr111 2022-02-26
I added a release note and split the PR in two parts. This one still focus on faster reduction operators. The other part about casting is #21123.