@seberg 2022-03-07
Lets put this in, looks like a nice speed improvement!  If the SIMD crows changes this, having the contiguous special case for comparison is probably good also.  Thanks @zephyr111.

(One day, I hope we can move these into the `ArrayMethod`'s `get_loop` function, but that is a different issue.)